+++ novatel_gps_driver/src/nodelets/novatel_gps_nodelet.cpp
@@ -134,7 +134,6 @@
 #include <boost/accumulators/statistics/min.hpp>
 #include <boost/accumulators/statistics/rolling_mean.hpp>
 #include <boost/accumulators/statistics/stats.hpp>
-#include <boost/accumulators/statistics/tail.hpp>
 #include <boost/accumulators/statistics/variance.hpp>
 #include <boost/circular_buffer.hpp>
 
@@ -177,11 +176,15 @@ namespace novatel_gps_driver
     NovatelGpsNodelet() :
       device_(""),
       connection_type_("serial"),
+      serial_baud_(115200),
       polling_period_(0.05),
       publish_gpgsa_(false),
       publish_gpgsv_(false),
+      publish_gphdt_(false),
       imu_rate_(100.0),
       imu_sample_rate_(-1),
+      span_frame_to_ros_frame_(false),
+      publish_clock_steering_(false),
       publish_imu_messages_(false),
       publish_novatel_positions_(false),
       publish_novatel_xyz_positions_(false),
@@ -197,7 +200,6 @@ namespace novatel_gps_driver
       publish_sync_diagnostic_(true),
       reconnect_delay_s_(0.5),
       use_binary_messages_(false),
-      span_frame_to_ros_frame_(false),
       connection_(NovatelGps::SERIAL),
       last_sync_(ros::TIME_MIN),
       rolling_offset_(stats::tag::rolling_window::window_size = 10),
@@ -215,7 +217,7 @@ namespace novatel_gps_driver
     {
     }
 
-    ~NovatelGpsNodelet()
+    ~NovatelGpsNodelet() override
     {
       gps_.Disconnect();
     }
@@ -224,7 +226,7 @@ namespace novatel_gps_driver
      * Init method reads parameters and sets up publishers and subscribers.
      * It does not connect to the device.
      */
-    void onInit()
+    void onInit() override
     {
       ros::NodeHandle &node = getNodeHandle();
       ros::NodeHandle &priv = getPrivateNodeHandle();
@@ -232,9 +234,10 @@ namespace novatel_gps_driver
       swri::param(priv, "device", device_, device_);
       swri::param(priv, "imu_rate", imu_rate_, imu_rate_);
       swri::param(priv, "imu_sample_rate", imu_sample_rate_, imu_sample_rate_);
-      swri::param(priv, "publish_clocksteering", publish_clock_steering_, false);
+      swri::param(priv, "publish_clocksteering", publish_clock_steering_, publish_clock_steering_);
       swri::param(priv, "publish_gpgsa", publish_gpgsa_, publish_gpgsa_);
       swri::param(priv, "publish_gpgsv", publish_gpgsv_, publish_gpgsv_);
+      swri::param(priv, "publish_gphdt", publish_gphdt_, publish_gphdt_);
       swri::param(priv, "publish_imu_messages", publish_imu_messages_, publish_imu_messages_);
       swri::param(priv, "publish_novatel_positions", publish_novatel_positions_, publish_novatel_positions_);
       swri::param(priv, "publish_novatel_xyz_positions", publish_novatel_xyz_positions_, publish_novatel_xyz_positions_);
@@ -255,7 +258,7 @@ namespace novatel_gps_driver
 
       swri::param(priv, "connection_type", connection_type_, connection_type_);
       connection_ = NovatelGps::ParseConnection(connection_type_);
-      swri::param(priv, "serial_baud", serial_baud_, 115200);
+      swri::param(priv, "serial_baud", serial_baud_, serial_baud_);
 
       swri::param(priv, "imu_frame_id", imu_frame_id_, std::string(""));
       swri::param(priv, "frame_id", frame_id_, std::string(""));
@@ -305,6 +308,11 @@ namespace novatel_gps_driver
         gpgsv_pub_ = swri::advertise<novatel_gps_msgs::Gpgsv>(node, "gpgsv", 100);
       }
 
+      if (publish_gphdt_)
+      {
+        gphdt_pub_ = swri::advertise<novatel_gps_msgs::Gphdt>(node,"gphdt", 100);
+      }
+
       if (publish_novatel_positions_)
       { 
         novatel_position_pub_ = swri::advertise<novatel_gps_msgs::NovatelPosition>(node, "bestpos", 100);
@@ -327,7 +335,7 @@ namespace novatel_gps_driver
 
       if (publish_novatel_heading2_)
       {
-        novatel_heading2_pub_ = swri::advertise<novatel_gps_msgs::NovatelHeading2>(node, "heading2", 100);
+	novatel_heading2_pub_ = swri::advertise<novatel_gps_msgs::NovatelHeading2>(node, "heading2", 100);
       }
 
       if (publish_novatel_dual_antenna_heading_)
@@ -426,7 +434,7 @@ namespace novatel_gps_driver
       }
       if (publish_novatel_dual_antenna_heading_)
       {
-        opts["dual_antenna_heading" + format_suffix] = polling_period_;
+        opts["dualantennaheading" + format_suffix] = polling_period_;
       }
       if (publish_gpgsa_)
       {
@@ -435,6 +443,10 @@ namespace novatel_gps_driver
       if (publish_gpgsv_)
       {
         opts["gpgsv"] = 1.0;
+      }
+      if (publish_gphdt_)
+      {
+        opts["gphdt"] = polling_period_;
       }
       if (publish_clock_steering_)
       {
@@ -554,6 +566,7 @@ namespace novatel_gps_driver
     double polling_period_;
     bool publish_gpgsa_;
     bool publish_gpgsv_;
+    bool publish_gphdt_;
     /// The rate at which IMU measurements will be published, in Hz
     double imu_rate_;
     /// How frequently the device samples the IMU, in Hz
@@ -594,6 +607,7 @@ namespace novatel_gps_driver
     ros::Publisher gpgga_pub_;
     ros::Publisher gpgsv_pub_;
     ros::Publisher gpgsa_pub_;
+    ros::Publisher gphdt_pub_;
     ros::Publisher gprmc_pub_;
     ros::Publisher range_pub_;
     ros::Publisher time_pub_;
@@ -647,7 +661,7 @@ namespace novatel_gps_driver
     bool resetService(novatel_gps_msgs::NovatelFRESET::Request& req,
                       novatel_gps_msgs::NovatelFRESET::Response& res)
     {
-      if (gps_.IsConnected() == false)
+      if (!gps_.IsConnected())
       {
         res.success = false;
       }
@@ -675,14 +689,9 @@ namespace novatel_gps_driver
      */
     void CheckDeviceForData()
     {
-      std::vector<novatel_gps_msgs::NovatelPositionPtr> position_msgs;
-      std::vector<novatel_gps_msgs::NovatelXYZPtr> xyz_position_msgs;
-      std::vector<novatel_gps_msgs::NovatelUtmPositionPtr> utm_msgs;
-      std::vector<novatel_gps_msgs::NovatelHeading2Ptr> heading2_msgs;
-      std::vector<novatel_gps_msgs::NovatelDualAntennaHeadingPtr> dual_antenna_heading_msgs;
       std::vector<gps_common::GPSFixPtr> fix_msgs;
+      std::vector<novatel_gps_msgs::NovatelPositionPtr> position_msgs;
       std::vector<novatel_gps_msgs::GpggaPtr> gpgga_msgs;
-      std::vector<novatel_gps_msgs::GprmcPtr> gprmc_msgs;
 
       // This call appears to block if the serial device is disconnected
       NovatelGps::ReadResult result = gps_.ProcessData();
@@ -717,15 +726,12 @@ namespace novatel_gps_driver
         gps_insufficient_data_warnings_++;
       }
 
-      // Read messages from the driver into the local message lists
-      gps_.GetGpggaMessages(gpgga_msgs);
-      gps_.GetGprmcMessages(gprmc_msgs);
-      gps_.GetNovatelPositions(position_msgs);
-      gps_.GetNovatelXYZPositions(xyz_position_msgs);
-      gps_.GetNovatelUtmPositions(utm_msgs);
+      // GPSFix messages are always published, and Gpgga and Position messages
+      // are used for generating some diagnostics.  Other message types will
+      // only be retrieved if we're configured to publish them.
       gps_.GetFixMessages(fix_msgs);
-      gps_.GetNovatelHeading2Messages(heading2_msgs);
-      gps_.GetNovatelDualAntennaHeadingMessages(dual_antenna_heading_msgs);
+      gps_.GetGpggaMessages(gpgga_msgs);
+      gps_.GetNovatelPositions(position_msgs);
 
       // Increment the measurement count by the number of messages we just
       // read
@@ -744,7 +750,7 @@ namespace novatel_gps_driver
       {
         if (msg->utc_seconds != 0)
         {
-          int64_t second = static_cast<int64_t>(swri_math_util::Round(msg->utc_seconds));
+          auto second = static_cast<int64_t>(swri_math_util::Round(msg->utc_seconds));
           double difference = std::fabs(msg->utc_seconds - second);
 
           if (difference < 0.02)
@@ -778,6 +784,8 @@ namespace novatel_gps_driver
           gpgga_pub_.publish(msg);
         }
 
+        std::vector<novatel_gps_msgs::GprmcPtr> gprmc_msgs;
+        gps_.GetGprmcMessages(gprmc_msgs);
         for (const auto& msg : gprmc_msgs)
         {
           msg->header.stamp += sync_offset;
@@ -810,6 +818,18 @@ namespace novatel_gps_driver
         }
       }
 
+      if (publish_gphdt_)
+      {
+        std::vector<novatel_gps_msgs::GphdtPtr> gphdt_msgs;
+        gps_.GetGphdtMessages(gphdt_msgs);
+        for (const auto& msg : gphdt_msgs)
+        {
+          msg->header.stamp = ros::Time::now();
+          msg->header.frame_id = frame_id_;
+          gphdt_pub_.publish(msg);
+        }
+      }
+
       if (publish_novatel_positions_)
       {
         for (const auto& msg : position_msgs)
@@ -822,6 +842,8 @@ namespace novatel_gps_driver
 
       if (publish_novatel_xyz_positions_)
       {
+        std::vector<novatel_gps_msgs::NovatelXYZPtr> xyz_position_msgs;
+        gps_.GetNovatelXYZPositions(xyz_position_msgs);
         for (const auto& msg : xyz_position_msgs)
         {
           msg->header.stamp += sync_offset;
@@ -832,6 +854,8 @@ namespace novatel_gps_driver
 
       if (publish_novatel_utm_positions_)
       {
+        std::vector<novatel_gps_msgs::NovatelUtmPositionPtr> utm_msgs;
+        gps_.GetNovatelUtmPositions(utm_msgs);
         for (const auto& msg : utm_msgs)
         {
           msg->header.stamp += sync_offset;
@@ -842,6 +866,8 @@ namespace novatel_gps_driver
 
       if (publish_novatel_heading2_)
       {
+        std::vector<novatel_gps_msgs::NovatelHeading2Ptr> heading2_msgs;
+        gps_.GetNovatelHeading2Messages(heading2_msgs);
         for (const auto& msg : heading2_msgs)
         {
           msg->header.stamp += sync_offset;
@@ -852,6 +878,8 @@ namespace novatel_gps_driver
 
       if (publish_novatel_dual_antenna_heading_)
       {
+        std::vector<novatel_gps_msgs::NovatelDualAntennaHeadingPtr> dual_antenna_heading_msgs;
+        gps_.GetNovatelDualAntennaHeadingMessages(dual_antenna_heading_msgs);
         for (const auto& msg : dual_antenna_heading_msgs)
         {
           msg->header.stamp += sync_offset;
@@ -1061,10 +1089,10 @@ namespace novatel_gps_driver
       int32_t synced_i = -1;  /// Index of last synced timesync msg
       int32_t synced_j = -1;  /// Index of last synced gps msg
       // Loop over sync times buffer
-      for (int32_t i = 0; i < sync_times_.size(); i++)
+      for (size_t i = 0; i < sync_times_.size(); i++)
       {
         // Loop over message times buffer
-        for (int32_t j = synced_j + 1; j < msg_times_.size(); j++)
+        for (size_t j = synced_j + 1; j < msg_times_.size(); j++)
         {
           // Offset is the difference between the sync time and message time
           double offset = (sync_times_[i] - msg_times_[j]).toSec();